chore: update nft recent
