+++ src/utils/canvas.ts
@@ -6,9 +6,10 @@ import {
   loadImage,
 } from "canvas"
 import { ChartJSNodeCanvas } from "chartjs-node-canvas"
-import { GuildMember } from "discord.js"
+import { GuildMember, MessageAttachment } from "discord.js"
 import { CircleleStats, RectangleStats } from "types/canvas"
-import { msgColors } from "./common"
+import { NFTCollection } from "types/community"
+import { msgColors, thumbnails } from "./common"
 import { SPACE } from "./constants"
 
 export function widthOf(ctx: CanvasRenderingContext2D, text: string): number {
@@ -373,4 +374,106 @@ export function handleTextOverflow(
     }
     return text + ellipsis
   }
+}
+
+// render canvas for nft list and nft recent
+export async function renderSupportedNFTList(collectionList: NFTCollection[]) {
+  const container: RectangleStats = {
+    x: {
+      from: 0,
+      to: 870,
+    },
+    y: {
+      from: 0,
+      to: 420,
+    },
+    w: 0,
+    h: 0,
+    pt: 0,
+    pl: 30,
+    radius: 30,
+    bgColor: "rgba(0, 0, 0, 0)", // transparent
+  }
+  container.w = container.x.to - container.x.from
+  container.h = container.y.to - container.y.from
+  const canvas = createCanvas(container.w, container.h)
+  const ctx = canvas.getContext("2d")
+
+  // background
+  ctx.save()
+  drawRectangle(ctx, container, container.bgColor)
+  ctx.clip()
+  ctx.restore()
+
+  const fixedCollectionNameHeight = 24
+  // const fixedChainNameHeight = 26
+  const iconConfig = {
+    w: 30,
+    h: 30,
+    mr: 20,
+  }
+  ctx.font = "27px Whitney"
+  let columnY = container.pt
+
+  collectionList = collectionList
+    .filter((col) => !!col.name)
+    .map((col) => {
+      col.image = col.image ? col.image : thumbnails.PROFILE
+      return col
+    })
+
+  const images: Record<string, Image> = await loadImages(
+    collectionList.map((col) => col.image)
+  )
+  collectionList.forEach((item, idx) => {
+    const colMaxWidth = 300
+    const symbolName = item.symbol?.toUpperCase()
+    const cName = item.name
+    const symbolNameWidth = widthOf(ctx, symbolName)
+
+    let collectionName: string
+    if (symbolNameWidth < colMaxWidth) {
+      const maxColNameWidth = colMaxWidth - symbolNameWidth
+      collectionName =
+        handleTextOverflow(ctx, cName, maxColNameWidth) +
+        ` (${item.symbol?.toUpperCase()})`
+    } else {
+      collectionName =
+        handleTextOverflow(ctx, cName, 80) +
+        ` (${handleTextOverflow(ctx, item.symbol?.toUpperCase(), 200)})`
+    }
+
+    const xStart = idx % 2 === 0 ? container.x.from : 440
+    const colConfig = {
+      x: xStart + iconConfig.w + iconConfig.mr,
+      y: container.pt,
+      mr: 10,
+      mb: 50,
+    }
+
+    // collection name
+    if (idx % 2 === 0) {
+      columnY +=
+        fixedCollectionNameHeight +
+        (iconConfig.h - fixedCollectionNameHeight) / 2 +
+        20
+    }
+
+    const conf: CircleleStats = {
+      x: xStart + 20,
+      y: columnY - 10,
+      radius: 20,
+    }
+    if (images[item.image]) {
+      drawCircleImage({ ctx, stats: conf, image: images[item.image] })
+    }
+
+    ctx.font = "semibold 27px Whitney"
+    ctx.fillStyle = "white"
+    ctx.fillText(collectionName, colConfig.x, columnY)
+
+    ctx.restore()
+  })
+
+  return new MessageAttachment(canvas.toBuffer(), "nftlist.png")
 }
